requirements.txt
streamlit
streamlit-authenticator
psycopg2-binary
